+++ .github/pull_request_template.md
@@ -22,15 +22,3 @@
 - [ ] 파일 혹은 폴더명 수정
 - [ ] 파일 혹은 폴더 삭제
 
-## 💬 공유사항 to 리뷰어
-
-<!--- 리뷰어가 중점적으로 봐줬으면 좋겠는 부분이 있으면 적어주세요. -->
-<!--- 논의해야할 부분이 있다면 적어주세요.-->
-<!--- ex) 메서드 XXX의 이름을 더 잘 짓고 싶은데 혹시 좋은 명칭이 있을까요? -->
-
-## ✅ PR Checklist
-
-PR이 다음 요구 사항을 충족하는지 확인하세요.
-
-- [ ] 커밋 메시지 컨벤션에 맞게 작성했습니다.
-- [ ] 변경 사항에 대한 테스트를 했습니다.(버그 수정/기능에 대한 테스트).
