+++ src/shared/components/ui/calendar.tsx
@@ -4,7 +4,7 @@ import * as React from "react"
 import { ChevronLeft, ChevronRight } from "lucide-react"
 import { DayPicker } from "react-day-picker"
 
-import { cn } from "@/lib/utils"
+import { cn } from "@/src/shared/lib/utils"
 import { buttonVariants } from "@/src/shared/components/ui/button"
 
 export type CalendarProps = React.ComponentProps<typeof DayPicker>
