+++ src/shared/components/ui/skeleton.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@/lib/utils"
+import { cn } from "@/src/shared/lib/utils"
 
 function Skeleton({
   className,
